# Speed up syntax highlighting